Ported over test for Patron.work_is_age_appropriate.
